+++ CHANGELOG.md
@@ -9,6 +9,7 @@ v1.2.2
 * #23: Upgrade Bootstrap to v3.0.2
 * #24: Upgrade Font Awesome to v4.0.3
 * #25: Upgrade AngularJS to v1.2.1
+* #26: Use angular.element() instead of $
 
 v1.2.1 (2013/10/18)